Moved utility function from pipeline to utils
