+++ pyaptamer/utils/_aptanet_utils.py
@@ -1,4 +1,8 @@
-DNA_BASES = list("ACGT")
+from itertools import product
+
+import numpy as np
+
+from pyaptamer.pseaac import PSeAAC
 
 
 def generate_kmer_vecs(aptamer_sequence, k=4):
@@ -21,9 +25,7 @@ def generate_kmer_vecs(aptamer_sequence, k=4):
         1D numpy array of normalized frequency vector for all possible k-mers from
         length 1 to k.
     """
-    from itertools import product
-
-    import numpy as np
+    DNA_BASES = list("ACGT")
 
     # Generate all possible k-mers from 1 to k
     all_kmers = []
@@ -49,3 +51,45 @@ def generate_kmer_vecs(aptamer_sequence, k=4):
     )
 
     return kmer_freq
+
+
+def pairs_to_features(X, k=4, pseaac_kwargs=None):
+    """
+    Convert a list of (aptamer_sequence, protein_sequence) pairs into feature vectors.
+
+    This function generates feature vectors for each (aptamer, protein) pair using:
+    - k-mer representation of the aptamer sequence
+    - Pseudo amino acid composition (PSeAAC) representation of the protein sequence
+
+    Parameters
+    ----------
+    X : list of tuple of str
+        A list where each element is a tuple `(aptamer_sequence, protein_sequence)`.
+        `aptamer_sequence` should be a string of nucleotides, and `protein_sequence`
+        should be a string of amino acids.
+
+    k : int, optional
+        The k-mer size used to generate the k-mer vector from the aptamer sequence.
+        Default is 4.
+
+    pseaac_kwargs : dict, optional
+        Optional keyword arguments to pass to the `PSeAAC` transformer.
+        If not provided, default parameters are used.
+
+    Returns
+    -------
+    np.ndarray
+        A 2D NumPy array where each row corresponds to the concatenated feature vector
+        for a given (aptamer, protein) pair.
+    """
+    pseaac_kwargs = {} if pseaac_kwargs is None else pseaac_kwargs
+    pseaac = PSeAAC(**pseaac_kwargs)
+
+    feats = []
+    for aptamer_seq, protein_seq in X:
+        kmer = generate_kmer_vecs(aptamer_seq, k=k)
+        pseaac_vec = np.asarray(pseaac.transform(protein_seq))
+        feats.append(np.concatenate([kmer, pseaac_vec]))
+
+    # Ensure float32 for PyTorch compatibility
+    return np.vstack(feats).astype(np.float32)
